Added some cleanup code to remove containers, messages and models adding during testing.

+++ trunk/Utils/KamanjaAppTester/src/main/scala/com/ligadata/kamanja/test/application/TestExecutor.scala
@@ -184,10 +184,16 @@ object TestExecutor {
           eventConsumer.shutdown
           errorConsumer.shutdown
 
+          if (!appManager.removeApplicationMetadata(app)) {
+            logger.error(s"*** Failed to remove metadata from application '${app.name}'")
+            throw new Exception(s"*** Failed to remove metadata from application '${app.name}'")
+          }
+
           if (KamanjaEnvironmentManager.isEmbedded) {
             KamanjaEnvironmentManager.stopServices
             TestUtils.deleteFile(EmbeddedConfiguration.storageDir)
           }
+
           logger.close
         }
       })
